+++ .env.example
@@ -1,0 +1,1 @@
+SEMANTIC_AI_BASE_DIR=/app/semantic_ai
