+++ .gitignore
@@ -172,3 +172,9 @@ cython_debug/
 
 # PyPI configuration file
 .pypirc
+
+# Project files
+*.db
+*.log
+.env
+venv/
