+++ CONTRIBUTING.md
@@ -1,0 +1,5 @@
+# Contributing
+
+1. Fork the repository and create your feature branch.
+2. Install dependencies via `pip install -r requirements.txt`.
+3. Run `python -m py_compile semantic_ai.py semantic_models.py` and `pytest -v` before submitting a pull request.
